Changes to get debug build working with gfortran 4.6 versions. These changes fix warnings relating to precision loss (for example when assigning a double precision result to a single precision variable.)
git-svn-id: https://repository.astro.ex.ac.uk/torus/trunk@3269 25464f21-6099-402c-b30a-640f1d40d38b

+++ torus/timedep_mod.F90
@@ -480,7 +480,7 @@ contains
     nFromMatter = nMonte
     nFromMatterThisThread = nFromMatter
 #ifdef MPI
-    nFromMatterThisThread = dble(nFromMatter) / dble(nThreadsGlobal)
+    nFromMatterThisThread = int(dble(nFromMatter) / dble(nThreadsGlobal))
 #endif
